requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
pytest>=7.4.0
requests>=2.31.0
redis>=4.5.0
celery>=5.3.0
watchdog>=3.0.0